+++ packages/image_picker/image_picker/android/src/main/java/io/flutter/plugins/imagepicker/ExifDataCopier.java
@@ -4,8 +4,8 @@
 
 package io.flutter.plugins.imagepicker;
 
-import android.media.ExifInterface;
 import android.util.Log;
+import androidx.exifinterface.media.ExifInterface;
 import java.util.Arrays;
 import java.util.List;