+++ packages/in_app_purchase/lib/src/store_kit_wrappers/sk_payment_transaction_wrappers.dart
@@ -98,6 +98,7 @@ enum SKPaymentTransactionStateWrapper {
 /// [SKPaymentTransaction](https://developer.apple.com/documentation/storekit/skpaymenttransaction?language=objc).
 @JsonSerializable(nullable: true)
 class SKPaymentTransactionWrapper {
+  /// Creates a new [SKPaymentTransactionWrapper] with the provided information.
   SKPaymentTransactionWrapper({
     @required this.payment,
     @required this.transactionState,
@@ -185,4 +186,10 @@ class SKPaymentTransactionWrapper {
 
   @override
   String toString() => _$SKPaymentTransactionWrapperToJson(this).toString();
+
+  /// The payload that is used to finish this transaction.
+  Map<String, String> toFinishMap() => {
+        "transactionIdentifier": this.transactionIdentifier,
+        "productIdentifier": this.payment?.productIdentifier,
+      };
 }