+++ packages/video_player/video_player/lib/src/closed_caption_file.dart
@@ -45,4 +45,13 @@ class Caption {
   /// The actual text that should appear on screen to be read between [start]
   /// and [end].
   final String text;
+
+  @override
+  String toString() {
+    return '$runtimeType('
+        'number: $number, '
+        'start: $start, '
+        'end: $end, '
+        'text: $text)';
+  }
 }