+++ script/incremental_build.sh
@@ -6,6 +6,12 @@ readonly REPO_DIR="$(dirname "$SCRIPT_DIR")"
 
 source "$SCRIPT_DIR/common.sh"
 
+if [ "$(expr substr $(uname -s) 1 5)" == "MINGW" ]; then
+  PUB=pub.bat
+else
+  PUB=pub
+fi
+
 # Plugins that deliberately use their own analysis_options.yaml.
 #
 # This list should only be deleted from, never added to. This only exists
@@ -14,9 +20,6 @@ source "$SCRIPT_DIR/common.sh"
 #
 # TODO(mklim): Remove everything from this list. https://github.com/flutter/flutter/issues/45440
 CUSTOM_ANALYSIS_PLUGINS=(
-  "in_app_purchase"
-  "camera"
-  "video_player/video_player_web"
 )
 # Comma-separated string of the list above
 readonly CUSTOM_FLAG=$(IFS=, ; echo "${CUSTOM_ANALYSIS_PLUGINS[*]}")
@@ -29,9 +32,14 @@ elif [[ "${ACTIONS[@]}" == "analyze" ]]; then
 fi
 
 BRANCH_NAME="${BRANCH_NAME:-"$(git rev-parse --abbrev-ref HEAD)"}"
+
+# This has to be turned into a list and then split out to the command line,
+# otherwise it gets treated as a single argument.
+PLUGIN_SHARDING=($PLUGIN_SHARDING)
+
 if [[ "${BRANCH_NAME}" == "master" ]]; then
   echo "Running for all packages"
-  (cd "$REPO_DIR" && pub global run flutter_plugin_tools "${ACTIONS[@]}" $PLUGIN_SHARDING)
+  (cd "$REPO_DIR" && $PUB global run flutter_plugin_tools "${ACTIONS[@]}" ${PLUGIN_SHARDING[@]})
 else
   # Sets CHANGED_PACKAGES
   check_changed_packages
@@ -39,11 +47,11 @@ else
   if [[ "$CHANGED_PACKAGES" == "" ]]; then
     echo "No changes detected in packages."
     echo "Running for all packages"
-    (cd "$REPO_DIR" && pub global run flutter_plugin_tools "${ACTIONS[@]}" $PLUGIN_SHARDING)
+    (cd "$REPO_DIR" && $PUB global run flutter_plugin_tools "${ACTIONS[@]}" ${PLUGIN_SHARDING[@]})
   else
     echo running "${ACTIONS[@]}"
-    (cd "$REPO_DIR" && pub global run flutter_plugin_tools "${ACTIONS[@]}" --plugins="$CHANGED_PACKAGES" $PLUGIN_SHARDING)
+    (cd "$REPO_DIR" && $PUB global run flutter_plugin_tools "${ACTIONS[@]}" --plugins="$CHANGED_PACKAGES" ${PLUGIN_SHARDING[@]})
     echo "Running version check for changed packages"
-    (cd "$REPO_DIR" && pub global run flutter_plugin_tools version-check --base_sha="$(get_branch_base_sha)")
+    (cd "$REPO_DIR" && $PUB global run flutter_plugin_tools version-check --base_sha="$(get_branch_base_sha)")
   fi
 fi