+++ packages/ios_platform_images/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.2+3
+
+* Remove no-op android folder in the example app.
+
+## 0.1.2+2
+
+* Post-v2 Android embedding cleanups.
+
 ## 0.1.2+1
 
 * Remove Android folder from `ios_platform_images`.