+++ packages/local_auth/CHANGELOG.md
@@ -1,3 +1,32 @@
+## 0.6.3+4
+
+* Update Dart SDK constraint in example.
+
+## 0.6.3+3
+
+* Update android compileSdkVersion to 29.
+
+## 0.6.3+2
+
+* Keep handling deprecated Android v1 classes for backward compatibility.
+
+## 0.6.3+1
+
+* Update package:e2e -> package:integration_test
+
+## 0.6.3
+
+* Increase upper range of `package:platform` constraint to allow 3.X versions.
+
+## 0.6.2+4
+
+* Update package:e2e reference to use the local version in the flutter/plugins
+  repository.
+
+## 0.6.2+3
+
+* Post-v2 Android embedding cleanup.
+
 ## 0.6.2+2
 
 * Update lower bound of dart dependency to 2.1.0.