+++ packages/path_provider/path_provider_macos/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.4+6
+
+* Remove unused `test` dependency.
+* Update Dart SDK constraint in example.
+
+## 0.0.4+5
+
+* Update license header.
+
+## 0.0.4+4
+
+* Remove no-op android folder in the example app.
+
 ## 0.0.4+3
 
 * Remove Android folder from `path_provider_macos`.