+++ packages/shared_preferences/shared_preferences_macos/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.1+11
+
+* Remove unused `test` dependency.
+* Update Dart SDK constraint in example.
+
+## 0.0.1+10
+
+* Remove iOS and Android folders from the example app.
+
 ## 0.0.1+9
 
 * Remove Android folder from `shared_preferences_macos`.