+++ packages/url_launcher/url_launcher_macos/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.0.1+9
+
+* Update Dart SDK constraint in example.
+
+# 0.0.1+8
+
+* Remove no-op android folder in the example app.
+
 # 0.0.1+7
 
 * Remove Android folder from url_launcher_web and url_launcher_macos.